Use unsigned long lng on all platforms for file size. Print human-readable filesize in ls_dir().

+++ src/api/api.c
@@ -1030,7 +1030,7 @@ void getVersion(const int *sock)
 void getDBstats(const int *sock)
 {
 	// Get file details
-	long int filesize = get_FTL_db_filesize();
+	unsigned long long int filesize = get_FTL_db_filesize();
 
 	char *prefix = calloc(2, sizeof(char));
 	if(prefix == NULL) return;
